+++ app/jobs/periodic/index.ts
@@ -7,12 +7,18 @@ import { Job, type JobHandler, JobManager } from '../../models';
 
 import { initHandlers as initAuthTokensHandlers } from './auth-tokens';
 import { initHandlers as initFrozenUsersHandlers } from './frozen-users';
+import { initHandlers as initTranslationUsageHandlers } from './translation-usage';
 
 const debugError = createDebug('freefeed:jobs:errors');
 
 export async function initHandlers(jobManager: JobManager, app: FreefeedApp) {
   await Promise.all(
-    [initAuthTokensHandlers, initFrozenUsersHandlers].map((h) => h(jobManager, app)),
+    [
+      // prettier-ignore
+      initAuthTokensHandlers,
+      initFrozenUsersHandlers,
+      initTranslationUsageHandlers,
+    ].map((h) => h(jobManager, app)),
   );
 }